Improve names

+++ src/problems/p25.rs
@@ -1,11 +1,11 @@
 pub fn p25() -> u64 {
-    smallest_fib_with_length(1000) as u64
+    smallest_fib_with_length(1000)
 }
 
-fn smallest_fib_with_length(len: u32) -> u32 {
+fn smallest_fib_with_length(len: u32) -> u64 {
     let sqrt_5 = 5f64.sqrt();
     let phi = (1f64 + sqrt_5) / 2f64;
-    ((len - 1) as f64 / phi.log(10f64) + sqrt_5.log(phi)).ceil() as u32
+    ((len - 1) as f64 / phi.log(10f64) + sqrt_5.log(phi)).ceil() as u64
 }
 
 #[cfg(test)]
